Removed BlockExplorer and AbstractBlockScheduler due single implementation.
