+++ app/src/test/java/org/hyperledger/besu/chainexport/RlpBlockExporterTest.java
@@ -58,7 +58,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-/** Tests for {@link BlockExporter}. */
+/** Tests for {@link RlpBlockExporter}. */
 @ExtendWith(MockitoExtension.class)
 public final class RlpBlockExporterTest {
